lc-support/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
langchain==0.3.0
langchain_community==0.3.0
langchain_openai==0.3.0
pinecone==5.3.0
openai==1.58.1
python-dotenv==1.0.1
pypdf==4.3.1
tiktoken==0.7.0
python-dotenv==1.0.1
uvicorn==0.23.2
dateutil
python-dateutil==2.8.2
streamlit
